moved bulk loading to driver management and fixed product styling reversion
