+++ client/src/components/SponsorDriverManagement.jsx
@@ -3,12 +3,12 @@ import 'bootstrap/dist/css/bootstrap.min.css';
 import SponsorNavbar from './SponsorNavbar';
 import { GenerateSalt } from './MiscellaneousParts/HashPass'; 
 import { updateDriverPoints } from "./MiscellaneousParts/ServerCall";
-//missing: getAllSponsorUsers, getAllDrivers, getAllSponsorUsers, addDriver, updateDriver, toggleDriverActivity/${driverID}, debugAllUsers
 
 export default function SponsorDriverManagement() {
     const [drivers, setDrivers] = useState([]);
     const [sponsorInfo, setSponsorInfo] = useState(null);
     const [loading, setLoading] = useState(true);
+    const [activeTab, setActiveTab] = useState('drivers'); // Add tab state
     const [showAddModal, setShowAddModal] = useState(false);
     const [showEditModal, setShowEditModal] = useState(false);
     const [editingDriver, setEditingDriver] = useState(null);
@@ -26,6 +26,12 @@ export default function SponsorDriverManagement() {
     const [search, setSearch] = useState(""); 
     const [showDebugModal, setShowDebugModal] = useState(false);
     const [debugData, setDebugData] = useState(null);
+
+    // Bulk upload states
+    const [bulkUploadFile, setBulkUploadFile] = useState(null);
+    const [bulkUploadResults, setBulkUploadResults] = useState(null);
+    const [bulkUploadLoading, setBulkUploadLoading] = useState(false);
+    const [isDragOver, setIsDragOver] = useState(false);
 
     const getUserInfo = () => {
         const userString = localStorage.getItem('user');
@@ -357,6 +363,232 @@ export default function SponsorDriverManagement() {
         );
     });
 
+    const handleBulkUpload = async (e) => {
+        e.preventDefault();
+        if (!bulkUploadFile) {
+            alert('Please select a file to upload.');
+            return;
+        }
+
+        setBulkUploadLoading(true);
+        setBulkUploadResults(null);
+
+        try {
+            const fileContent = await readFileContent(bulkUploadFile);
+            const lines = fileContent.split('\n').map(line => line.trim()).filter(line => line);
+            
+            const results = {
+                totalLines: lines.length,
+                processed: 0,
+                errors: [],
+                success: {
+                    drivers: 0,
+                    sponsors: 0
+                }
+            };
+
+            for (let i = 0; i < lines.length; i++) {
+                const lineNumber = i + 1;
+                const line = lines[i];
+                
+                try {
+                    const result = await processUploadLine(line, lineNumber);
+                    
+                    if (result.success) {
+                        results.success[result.type]++;
+                    } else {
+                        results.errors.push({
+                            line: lineNumber,
+                            content: line,
+                            error: result.error
+                        });
+                    }
+                } catch (error) {
+                    results.errors.push({
+                        line: lineNumber,
+                        content: line,
+                        error: `Unexpected error: ${error.message}`
+                    });
+                }
+                
+                results.processed++;
+            }
+
+            setBulkUploadResults(results);
+            
+            // Refresh data after bulk upload
+            if (sponsorInfo) {
+                fetchDriversForSponsor(sponsorInfo.SponsorID);
+            }
+            
+        } catch (error) {
+            console.error('Error processing bulk upload:', error);
+            alert(`Error processing file: ${error.message}`);
+        } finally {
+            setBulkUploadLoading(false);
+        }
+    };
+
+    const processUploadLine = async (line, lineNumber) => {
+        const parts = line.split('|');
+        
+        if (parts.length < 2) {
+            return { success: false, error: 'Invalid format: Not enough fields (minimum 2 required)' };
+        }
+
+        const type = parts[0].trim().toUpperCase();
+        
+        switch (type) {
+            case 'D':
+                return await processDriverRecord(parts);
+            case 'S':
+                return await processSponsorRecord(parts);
+            default:
+                return { success: false, error: `Invalid type '${type}'. Must be D or S.` };
+        }
+    };
+
+    const processDriverRecord = async (parts) => {
+        if (parts.length !== 5) {
+            return { success: false, error: 'Driver record must have exactly 4 fields: D||first name|last name|email' };
+        }
+
+        const [, organizationName, firstName, lastName, email] = parts.map(p => p.trim());
+        
+        if (!firstName || !lastName || !email) {
+            return { success: false, error: 'All driver fields are required and cannot be empty' };
+        }
+
+        if(organizationName){
+            return { success: false, error: 'Organization name cannot be provided by sponsor users.'}
+        }
+
+        // Validate email format
+        const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
+        if (!emailRegex.test(email)) {
+            return { success: false, error: 'Invalid email format' };
+        }
+
+        try {
+            const salt = GenerateSalt();
+            const driverData = {
+                FirstName: firstName,
+                LastName: lastName,
+                Email: email,
+                Password: 'DefaultPassword123!',
+                SponsorID: sponsorInfo.SponsorID,
+                UserType: 1,
+                PasswordSalt: salt
+            };
+
+            const queryString = new URLSearchParams(driverData).toString();
+            const response = await fetch(`https://63iutwxr2owp72oyfbetwyluaq0wakdm.lambda-url.us-east-1.on.aws/driverAPI/addDriver?${queryString}`, {
+                method: 'POST',
+                headers: {
+                    'Content-Type': 'application/json',
+                }
+            });
+
+            if (response.ok) {
+                return { success: true, type: 'drivers' };
+            } else {
+                const errorText = await response.text();
+                return { success: false, error: `Failed to create driver: ${errorText}` };
+            }
+        } catch (error) {
+            return { success: false, error: `Network error creating driver: ${error.message}` };
+        }
+    };
+
+    const processSponsorRecord = async (parts) => {
+        if (parts.length !== 5) {
+            return { success: false, error: 'Sponsor record must have exactly 4 fields: S||first name|last name|email' };
+        }
+
+        const [, organizationName, firstName, lastName, email] = parts.map(p => p.trim());
+        
+        if (!firstName || !lastName || !email) {
+            return { success: false, error: 'All sponsor fields are required and cannot be empty' };
+        }
+
+        if(organizationName){
+            return { success: false, error: 'Organization name cannot be provided by sponsor users.'}
+        }
+
+        // Validate email format
+        const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
+        if (!emailRegex.test(email)) {
+            return { success: false, error: 'Invalid email format' };
+        }
+
+        try {
+            const salt = GenerateSalt();
+            const sponsorData = {
+                FirstName: firstName,
+                LastName: lastName,
+                Email: email,
+                Password: 'DefaultPassword123!',
+                SponsorID: sponsorInfo.SponsorID,
+                UserType: 2,
+                PasswordSalt: salt
+            };
+
+            const queryString = new URLSearchParams(sponsorData).toString();
+            const response = await fetch(`https://63iutwxr2owp72oyfbetwyluaq0wakdm.lambda-url.us-east-1.on.aws/sponsorAPI/addSponsorUser?${queryString}`, {
+                method: 'POST',
+                headers: {
+                    'Content-Type': 'application/json',
+                },
+            });
+
+            if (response.ok) {
+                return { success: true, type: 'sponsors' };
+            } else {
+                const errorText = await response.text();
+                return { success: false, error: `Failed to create sponsor: ${errorText}` };
+            }
+        } catch (error) {
+            return { success: false, error: `Network error creating sponsor: ${error.message}` };
+        }
+    };
+
+    const readFileContent = (file) => {
+        return new Promise((resolve, reject) => {
+            const reader = new FileReader();
+            reader.onload = (e) => resolve(e.target.result);
+            reader.onerror = (e) => reject(new Error('Failed to read file'));
+            reader.readAsText(file);
+        });
+    };
+
+    const handleFileSelect = (file) => {
+        if (file && (file.type === 'text/plain' || file.name.endsWith('.txt'))) {
+            setBulkUploadFile(file);
+            setBulkUploadResults(null);
+        } else {
+            alert('Please select a valid text file (.txt)');
+        }
+    };
+
+    const handleDragOver = (e) => {
+        e.preventDefault();
+        setIsDragOver(true);
+    };
+
+    const handleDragLeave = (e) => {
+        e.preventDefault();
+        setIsDragOver(false);
+    };
+
+    const handleDrop = (e) => {
+        e.preventDefault();
+        setIsDragOver(false);
+        const files = Array.from(e.dataTransfer.files);
+        if (files.length > 0) {
+            handleFileSelect(files[0]);
+        }
+    };
+
     useEffect(() => {
         fetchSponsorInfo();
     }, []);
@@ -395,108 +627,378 @@ export default function SponsorDriverManagement() {
             <SponsorNavbar />
             <div className="container mt-4">
                 <div className="d-flex justify-content-between align-items-center mb-4">
-                    <h2>Driver Management - All Sponsors</h2>
-                    <div>
+                    <h2>Driver Management</h2>
+                    {activeTab === 'drivers' && (
+                        <div>
+                            <button 
+                                className="btn btn-outline-info me-2"
+                                onClick={handleDebugAllUsers}
+                            >
+                                <i className="fas fa-bug me-2"></i>
+                                Debug All Users
+                            </button>
+                            <button 
+                                className="btn btn-primary"
+                                onClick={() => setShowAddModal(true)}
+                            >
+                                <i className="fas fa-plus me-2"></i>
+                                Add New Driver
+                            </button>
+                        </div>
+                    )}
+                </div>
+
+                {/* Tab Navigation */}
+                <ul className="nav nav-tabs mb-3">
+                    <li className="nav-item">
                         <button 
-                            className="btn btn-outline-info me-2"
-                            onClick={handleDebugAllUsers}
+                            className={`nav-link ${activeTab === 'drivers' ? 'active' : ''}`}
+                            onClick={() => setActiveTab('drivers')}
                         >
-                            <i className="fas fa-bug me-2"></i>
-                            Debug All Users
+                            <i className="fas fa-users me-2"></i>
+                            Driver Management
                         </button>
+                    </li>
+                    <li className="nav-item">
                         <button 
-                            className="btn btn-primary"
-                            onClick={() => setShowAddModal(true)}
+                            className={`nav-link ${activeTab === 'bulk' ? 'active' : ''}`}
+                            onClick={() => setActiveTab('bulk')}
                         >
-                            <i className="fas fa-plus me-2"></i>
-                            Add New Driver
+                            <i className="fas fa-upload me-2"></i>
+                            Bulk Load
                         </button>
-                    </div>
-                </div>
+                    </li>
+                </ul>
 
-                {/* Search input */}
-                <div className="mb-3">
-                    <input
-                        type="text"
-                        className="form-control"
-                        placeholder="Search drivers by name, email, or ID..."
-                        value={search}
-                        onChange={e => setSearch(e.target.value)}
-                    />
-                </div>
+                {/* Driver Management Tab */}
+                {activeTab === 'drivers' && (
+                    <>
+                        {/* Search input */}
+                        <div className="mb-3">
+                            <input
+                                type="text"
+                                className="form-control"
+                                placeholder="Search drivers by name, email, or ID..."
+                                value={search}
+                                onChange={e => setSearch(e.target.value)}
+                            />
+                        </div>
 
-                <div className="card mb-3">
-                    <div className="card-body">
-                        <h5 className="card-title">Sponsor Information</h5>
-                        <p className="card-text">
-                            <strong>Sponsor ID:</strong> {sponsorInfo.SponsorID} <br />
-                            <strong>Total Drivers:</strong> {drivers.length}
-                        </p>
-                    </div>
-                </div>
+                        <div className="card mb-3">
+                            <div className="card-body">
+                                <h5 className="card-title">Sponsor Information</h5>
+                                <p className="card-text">
+                                    <strong>Sponsor ID:</strong> {sponsorInfo?.SponsorID} <br />
+                                    <strong>Total Drivers:</strong> {drivers.length}
+                                </p>
+                            </div>
+                        </div>
 
-                {filteredDrivers.length === 0 ? (
-                    <div className="alert alert-info">
-                        <h5>No Drivers Found</h5>
-                        <p>
-                            {search
-                                ? "No drivers match your search."
-                                : "You don't have any drivers in your program yet. Click \"Add New Driver\" to get started."
-                            }
-                        </p>
-                    </div>
-                ) : (
-                    <div className="table-responsive">
-                        <table className="table table-striped">
-                            <thead className="table-dark">
-                                <tr>
-                                    <th>Name</th>
-                                    <th>Email</th>
-                                    <th>User ID</th>
-                                    <th>Points</th>
-                                    <th>Actions</th>
-                                </tr>
-                            </thead>
-                            <tbody>
-                                {filteredDrivers.map((driver) => (
-                                    <tr key={driver.uniqueKey || driver.DriverID || `user-${driver.UserID}`} className={driver.ActiveAccount === 0 ? 'table-secondary' : ''}>
-                                        <td>
-                                            {driver.FirstName} {driver.LastName}
-                                            {driver.ActiveAccount === 0 && (
-                                                <span className="badge bg-danger ms-2">Inactive</span>
-                                            )}
-                                        </td>
-                                        <td>{driver.Email}</td>
-                                        <td>{driver.UserID}</td>
-                                        <td>{driver.Points}</td>
-                                        <td>
-                                            <button 
-                                                className="btn btn-sm btn-outline-primary me-2"
-                                                onClick={() => handleEditDriver(driver)}
+                        {filteredDrivers.length === 0 ? (
+                            <div className="alert alert-info">
+                                <h5>No Drivers Found</h5>
+                                <p>
+                                    {search
+                                        ? "No drivers match your search."
+                                        : "You don't have any drivers in your program yet. Click \"Add New Driver\" to get started."
+                                    }
+                                </p>
+                            </div>
+                        ) : (
+                            <div className="table-responsive">
+                                <table className="table table-striped">
+                                    <thead className="table-dark">
+                                        <tr>
+                                            <th>Name</th>
+                                            <th>Email</th>
+                                            <th>User ID</th>
+                                            <th>Points</th>
+                                            <th>Actions</th>
+                                        </tr>
+                                    </thead>
+                                    <tbody>
+                                        {filteredDrivers.map((driver) => (
+                                            <tr key={driver.uniqueKey || driver.DriverID || `user-${driver.UserID}`} className={driver.ActiveAccount === 0 ? 'table-secondary' : ''}>
+                                                <td>
+                                                    {driver.FirstName} {driver.LastName}
+                                                    {driver.ActiveAccount === 0 && (
+                                                        <span className="badge bg-danger ms-2">Inactive</span>
+                                                    )}
+                                                </td>
+                                                <td>{driver.Email}</td>
+                                                <td>{driver.UserID}</td>
+                                                <td>{driver.Points}</td>
+                                                <td>
+                                                    <button 
+                                                        className="btn btn-sm btn-outline-primary me-2"
+                                                        onClick={() => handleEditDriver(driver)}
+                                                    >
+                                                        <i className="fas fa-edit me-1"></i>
+                                                        Edit
+                                                    </button>
+                                                    <button 
+                                                        className="btn btn-sm btn-outline-info me-2"
+                                                        onClick={() => handleManagePointsClick(driver)}
+                                                    >
+                                                        <i className="fas fa-coins me-1"></i>
+                                                        Points
+                                                    </button>
+                                                    <button 
+                                                        className={`btn btn-sm ${driver.ActiveAccount === 1 ? 'btn-outline-warning' : 'btn-outline-success'}`}
+                                                        onClick={() => handleRemoveDriver(driver.DriverID || driver.UserID, `${driver.FirstName} ${driver.LastName}`, driver.ActiveAccount === 1)}
+                                                    >
+                                                        <i className={`fas ${driver.ActiveAccount === 1 ? 'fa-ban' : 'fa-check'} me-1`}></i>
+                                                        {driver.ActiveAccount === 1 ? 'Deactivate' : 'Reactivate'}
+                                                    </button>
+                                                </td>
+                                            </tr>
+                                        ))}
+                                    </tbody>
+                                </table>
+                            </div>
+                        )}
+                    </>
+                )}
+
+                {/* Bulk Load Tab */}
+                {activeTab === 'bulk' && (
+                    <>
+                        <div className="row">
+                            <div className="col-md-8">
+                                <div className="card">
+                                    <div className="card-body">
+                                        <h5 className="card-title">
+                                            <i className="fas fa-upload me-2"></i>
+                                            Bulk Load Users
+                                        </h5>
+                                        <p className="card-text">
+                                            Upload a pipe-delimited text file to create multiple drivers and sponsors at once.
+                                        </p>
+
+                                        <form onSubmit={handleBulkUpload}>
+                                            <div 
+                                                className={`border rounded p-4 mb-3 text-center ${isDragOver ? 'border-primary bg-light' : 'border-dashed'}`}
+                                                onDragOver={handleDragOver}
+                                                onDragLeave={handleDragLeave}
+                                                onDrop={handleDrop}
+                                                style={{ 
+                                                    borderStyle: isDragOver ? 'solid' : 'dashed',
+                                                    minHeight: '120px',
+                                                    display: 'flex',
+                                                    alignItems: 'center',
+                                                    justifyContent: 'center',
+                                                    flexDirection: 'column'
+                                                }}
                                             >
-                                                <i className="fas fa-edit me-1"></i>
-                                                Edit
-                                            </button>
-                                            <button 
-                                                className="btn btn-sm btn-outline-info me-2"
-                                                onClick={() => handleManagePointsClick(driver)}
-                                            >
-                                                <i className="fas fa-coins me-1"></i>
-                                                Points
-                                            </button>
-                                            <button 
-                                                className={`btn btn-sm ${driver.ActiveAccount === 1 ? 'btn-outline-warning' : 'btn-outline-success'}`}
-                                                onClick={() => handleRemoveDriver(driver.DriverID || driver.UserID, `${driver.FirstName} ${driver.LastName}`, driver.ActiveAccount === 1)}
-                                            >
-                                                <i className={`fas ${driver.ActiveAccount === 1 ? 'fa-ban' : 'fa-check'} me-1`}></i>
-                                                {driver.ActiveAccount === 1 ? 'Deactivate' : 'Reactivate'}
-                                            </button>
-                                        </td>
-                                    </tr>
-                                ))}
-                            </tbody>
-                        </table>
-                    </div>
+                                                {bulkUploadFile ? (
+                                                    <div>
+                                                        <i className="fas fa-file-alt fa-2x text-success mb-2"></i>
+                                                        <p className="mb-0">
+                                                            <strong>{bulkUploadFile.name}</strong>
+                                                        </p>
+                                                        <small className="text-muted">
+                                                            {(bulkUploadFile.size / 1024).toFixed(2)} KB
+                                                        </small>
+                                                    </div>
+                                                ) : (
+                                                    <div>
+                                                        <i className="fas fa-cloud-upload-alt fa-2x text-muted mb-2"></i>
+                                                        <p className="mb-2">
+                                                            Drag and drop your text file here, or click to browse
+                                                        </p>
+                                                        <input
+                                                            type="file"
+                                                            className="form-control"
+                                                            accept=".txt,text/plain"
+                                                            onChange={(e) => handleFileSelect(e.target.files[0])}
+                                                            style={{ maxWidth: '300px', margin: '0 auto' }}
+                                                        />
+                                                    </div>
+                                                )}
+                                            </div>
+
+                                            <div className="d-flex justify-content-between">
+                                                <button
+                                                    type="button"
+                                                    className="btn btn-outline-secondary"
+                                                    onClick={() => {
+                                                        setBulkUploadFile(null);
+                                                        setBulkUploadResults(null);
+                                                    }}
+                                                    disabled={!bulkUploadFile}
+                                                >
+                                                    Clear File
+                                                </button>
+                                                <button
+                                                    type="submit"
+                                                    className="btn btn-primary"
+                                                    disabled={!bulkUploadFile || bulkUploadLoading}
+                                                >
+                                                    {bulkUploadLoading ? (
+                                                        <>
+                                                            <span className="spinner-border spinner-border-sm me-2" role="status"></span>
+                                                            Processing...
+                                                        </>
+                                                    ) : (
+                                                        <>
+                                                            <i className="fas fa-upload me-2"></i>
+                                                            Upload and Process
+                                                        </>
+                                                    )}
+                                                </button>
+                                            </div>
+                                        </form>
+
+                                        {/* Results Display */}
+                                        {bulkUploadResults && (
+                                            <div className="mt-4">
+                                                <hr />
+                                                <h6>Upload Results</h6>
+                                                <div className="row mb-3">
+                                                    <div className="col-md-3">
+                                                        <div className="card bg-primary text-white">
+                                                            <div className="card-body text-center">
+                                                                <h5>{bulkUploadResults.totalLines}</h5>
+                                                                <small>Total Lines</small>
+                                                            </div>
+                                                        </div>
+                                                    </div>
+                                                    <div className="col-md-3">
+                                                        <div className="card bg-success text-white">
+                                                            <div className="card-body text-center">
+                                                                <h5>
+                                                                    {bulkUploadResults.success.drivers + 
+                                                                     bulkUploadResults.success.sponsors}
+                                                                </h5>
+                                                                <small>Successful</small>
+                                                            </div>
+                                                        </div>
+                                                    </div>
+                                                    <div className="col-md-3">
+                                                        <div className="card bg-danger text-white">
+                                                            <div className="card-body text-center">
+                                                                <h5>{bulkUploadResults.errors.length}</h5>
+                                                                <small>Errors</small>
+                                                            </div>
+                                                        </div>
+                                                    </div>
+                                                    <div className="col-md-3">
+                                                        <div className="card bg-info text-white">
+                                                            <div className="card-body text-center">
+                                                                <h5>{bulkUploadResults.processed}</h5>
+                                                                <small>Processed</small>
+                                                            </div>
+                                                        </div>
+                                                    </div>
+                                                </div>
+
+                                                <div className="row mb-3">
+                                                    <div className="col-md-6">
+                                                        <div className="text-center">
+                                                            <i className="fas fa-car text-primary fa-2x"></i>
+                                                            <h6 className="mt-2">Drivers</h6>
+                                                            <span className="badge bg-primary">{bulkUploadResults.success.drivers}</span>
+                                                        </div>
+                                                    </div>
+                                                    <div className="col-md-6">
+                                                        <div className="text-center">
+                                                            <i className="fas fa-handshake text-success fa-2x"></i>
+                                                            <h6 className="mt-2">Sponsors</h6>
+                                                            <span className="badge bg-success">{bulkUploadResults.success.sponsors}</span>
+                                                        </div>
+                                                    </div>
+                                                </div>
+
+                                                {/* Error Details */}
+                                                {bulkUploadResults.errors.length > 0 && (
+                                                    <div className="mt-3">
+                                                        <h6 className="text-danger">
+                                                            <i className="fas fa-exclamation-triangle me-2"></i>
+                                                            Errors ({bulkUploadResults.errors.length})
+                                                        </h6>
+                                                        <div className="table-responsive" style={{ maxHeight: '300px', overflowY: 'auto' }}>
+                                                            <table className="table table-sm table-striped">
+                                                                <thead className="table-dark">
+                                                                    <tr>
+                                                                        <th>Line</th>
+                                                                        <th>Content</th>
+                                                                        <th>Error</th>
+                                                                    </tr>
+                                                                </thead>
+                                                                <tbody>
+                                                                    {bulkUploadResults.errors.map((error, index) => (
+                                                                        <tr key={index}>
+                                                                            <td>{error.line}</td>
+                                                                            <td>
+                                                                                <code style={{ fontSize: '0.8em' }}>
+                                                                                    {error.content.length > 50 
+                                                                                        ? error.content.substring(0, 50) + '...' 
+                                                                                        : error.content}
+                                                                                </code>
+                                                                            </td>
+                                                                            <td className="text-danger" style={{ fontSize: '0.9em' }}>
+                                                                                {error.error}
+                                                                            </td>
+                                                                        </tr>
+                                                                    ))}
+                                                                </tbody>
+                                                            </table>
+                                                        </div>
+                                                    </div>
+                                                )}
+                                            </div>
+                                        )}
+                                    </div>
+                                </div>
+                            </div>
+
+                            <div className="col-md-4">
+                                <div className="card">
+                                    <div className="card-body">
+                                        <h6 className="card-title">
+                                            <i className="fas fa-info-circle me-2"></i>
+                                            File Format Instructions
+                                        </h6>
+                                        <div className="mb-3">
+                                            <h6>Record Types:</h6>
+                                            <ul className="list-unstyled">
+                                                <li><code>D</code> - Driver</li>
+                                                <li><code>S</code> - Sponsor User</li>
+                                            </ul>
+                                        </div>
+
+                                        <div className="mb-3">
+                                            <h6>Format Examples:</h6>
+                                            <div className="bg-light p-2 rounded">
+                                                <code style={{ fontSize: '0.8em' }}>
+                                                    D||Joe|Driver|joe@email.com<br />
+                                                    S||Jill|Sponsor|jill@mail.com
+                                                </code>
+                                            </div>
+                                        </div>
+
+                                        <div className="mb-3">
+                                            <h6>Rules:</h6>
+                                            <ul style={{ fontSize: '0.9em' }}>
+                                                <li>Use pipe (|) as delimiter</li>
+                                                <li>Leave organization field empty (second field)</li>
+                                                <li>No pipes allowed in field data</li>
+                                                <li>Email addresses must be valid format</li>
+                                                <li>Default password: "DefaultPassword123!"</li>
+                                                <li>All users will be assigned to your organization</li>
+                                                <li>Errors are skipped, processing continues</li>
+                                            </ul>
+                                        </div>
+
+                                        <div className="alert alert-warning" style={{ fontSize: '0.8em' }}>
+                                            <strong>Note:</strong> Sponsors cannot create organizations through bulk upload. 
+                                            All users will be automatically assigned to your sponsor organization.
+                                        </div>
+                                    </div>
+                                </div>
+                            </div>
+                        </div>
+                    </>
                 )}
 
                 {/* Add Driver Modal */}
