+++ client/src/components/SponsorNavbar.jsx
@@ -93,9 +93,6 @@ export default function SponsorNavbar() {
                             <li className="nav-item">
                                 <Link className="nav-link" to="/sponsorauditview">Audit</Link>
                             </li>
-                            <li className="nav-item">
-                                <Link className="nav-link" to="/sponsorbulkupload">Bulk Upload</Link>
-                            </li>
                         </ul>
                         <div className="d-flex align-items-center">
                             {isAdminImpostor && (
